html: Expose sanitize API on Html
