+++ crates/ruma-html/tests/it/html.rs
@@ -1,0 +1,1 @@
+mod sanitize;
