+++ crates/ruma-html/tests/it/main.rs
@@ -1,0 +1,2 @@
+mod helpers;
+mod html;
